moved user folder into models folder
