+++ __test__/user-actions.spec.js
@@ -1,5 +1,5 @@
 const makeFakeUser = require('./fixtures/user')
-const makeUser = require('../src/domain/user')
+const makeUser = require('../src/domain/models/user')
 const makeUserDb = require('./stubs/userdb-stub')
 const makeUserActions = require('../src/domain/services/user-actions')
